LayoutTests/touchadjustment/block-testing-expected.txt

Some text. Some text. A link
Test touch-adjustment on a block where we also touch the inner text and outside link.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS adjustedNode.id is "div1"
PASS successfullyParsed is true

TEST COMPLETE
